hw3/requirements.txt
pytest
pytest-xdist
pylint
requests
allure-pytest